IDE-3110 update logic for getting the blade jar path
We need to get the correct path to the blade jar, here is the logic
First see if we have a instance (workbench) copy of blade cli jar, that means
that the developer has intentially updated their blade jar to a newer version than
is shipped with the project.core bundle.  The local instance copy will be in the
plugin state area.

If there is no instance copy of blade cli jar, then fallback to use the one that
is in the bundle itself.