IDE-3972 inject the UpgradePlanner service instead of added to getter API
The UpgradePlanner is available as a osgi service.  So from a class
such as FindUpgradeProblems which is also an DS component
we can simply inject this service instead of polluting the API
of the UpgradeTaskStep to 'get' the upgradePlanner service.

UpgradeTaskStep is a low-level Api class not a high level, so we
shouldn't add a getter to a low level service to get a high level
service